continued updating translation system

+++ components/SNA/CrowdTangleSna/Results/CrowdTangleSnaResults.jsx
@@ -16,9 +16,9 @@ import UrlList from "../../Components/UrlList";
 import Count from "../Components/Count";
 import { snaTypeCleaned } from "../../../../redux/slices/tools/snaTypeSlice";
 import { csvSnaStateCleaned } from "../../../../redux/slices/tools/crowdTangleSnaSlice";
+import { i18nLoadNamespace } from "../../../shared/languages/i18nLoadNamespace";
 
-const tsv = "/components/NavItems/tools/TwitterSna.tsv";
-const tsv2 = "/components/NavItems/tools/CrowdTangle.tsv";
+
 const PlotTimeLine = dynamic(import("../Components/PlotTimeLine"), {
 	ssr: false,
 });
@@ -33,6 +33,11 @@ const CloudChart = dynamic(import("../Components/CloudChart"), { ssr: false });
 
 
 export default function CrowdTangleSnaResults(props) {
+
+
+	const tsv = "/components/NavItems/tools/TwitterSna.tsv";
+	const tsv2 = "/components/NavItems/tools/CrowdTangle.tsv";
+
 	const classes = useMyStyles();
 	const dispatch = useDispatch();
 	const keyword = useLoadLanguage(tsv);
@@ -45,10 +50,14 @@ export default function CrowdTangleSnaResults(props) {
 	useEffect(() => {
 
 		return () => {
-			props.workers.current.cloudWorker.terminate()
-			props.workers.current.hashtagWorker.terminate()
-			props.workers.current.pieChartsWorker.terminate()
-			props.workers.current.timelineWorker.terminate()
+
+			if(props.workers.current) {
+				props.workers.current.cloudWorker.terminate()
+				props.workers.current.hashtagWorker.terminate()
+				props.workers.current.pieChartsWorker.terminate()
+				props.workers.current.timelineWorker.terminate()
+			}
+			
 		}
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, []);
@@ -105,14 +114,14 @@ export default function CrowdTangleSnaResults(props) {
 
 									{!collapsed &&
 										<Grid item>
-											<IconButton onClick={props.functionNodes} onClick={onClickCollapseIndex}>
+											<IconButton onClick={onClickCollapseIndex}>
 												<ArrowBackIosIcon style={{ marginRight: "-5px", color: "white"}} />
 											</IconButton>
 										</Grid>
 									}
 									{collapsed &&
 										<Grid item>
-											<IconButton onClick={props.functionNodes} onClick={onClickCollapseIndex}>
+											<IconButton onClick={onClickCollapseIndex}>
 												<ArrowForwardIosIcon style={{color: "white"}} />
 											</IconButton>
 										</Grid>
@@ -333,7 +342,7 @@ export default function CrowdTangleSnaResults(props) {
 							{
 								props.result.urls &&
 								<UrlList result={props.result} title_message={'ct_sna_result_url_in_posts'}
-									tooltip_message={'twittersna_result_submit_twitter_sna'} downloadable={false} topic={"this topic"}/>
+									tooltip_message={'twittersna_result_submit_twitter_sna'} downloadable={false} topic={"this topic"} keyword={keyword2}/>
 							}
 						</div>
 					}
